+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/UnlinkConnectionRequest.cs
@@ -11,12 +11,13 @@
 namespace CodatSyncExpenses.Models.Operations
 {
     using CodatSyncExpenses.Utils;
+    using System.Collections.Generic;
     
     public class UnlinkConnectionRequest
     {
 
         [SpeakeasyMetadata("request:mediaType=application/json")]
-        public UnlinkConnectionUpdateConnection? RequestBody { get; set; }
+        public Dictionary<string, object>? RequestBody { get; set; }
 
         /// <summary>
         /// Unique identifier for a company.