+++ sync-for-expenses/RELEASES.md
@@ -101,3 +101,13 @@ Based on:
 - [csharp v3.1.0] sync-for-expenses
 ### Releases
 - [NuGet v3.1.0] https://www.nuget.org/packages/Codat.Sync.Expenses/3.1.0 - sync-for-expenses
+
+## 2023-10-09 01:27:05
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.96.1 (2.150.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.1.1] sync-for-expenses
+### Releases
+- [NuGet v3.1.1] https://www.nuget.org/packages/Codat.Sync.Expenses/3.1.1 - sync-for-expenses